StandaloneLoader - convert to functional, split out StandaloneLayout
